+++ udata/features/territories/api.py
@@ -6,7 +6,7 @@ from udata.features.territories import check_for_territories
 
 suggest_parser = api.parser()
 suggest_parser.add_argument(
-    'q', type=str, help='The string to autocomplete/suggest',
+    'q', type=unicode, help='The string to autocomplete/suggest',
     location='args', required=True)
 suggest_parser.add_argument(
     'size', type=int, help='The maximum result size',
@@ -18,7 +18,7 @@ class SuggestTerritoriesAPI(API):
     @api.doc(id='suggest_territory', parser=suggest_parser)
     def get(self):
         args = suggest_parser.parse_args()
-        territories = check_for_territories(args['q'].decode('utf-8'))
+        territories = check_for_territories(args['q'])
         if args['size']:
             territories = territories[:args['size']]
         return [{