+++ udata/tests/api/test_tags_api.py
@@ -32,6 +32,28 @@ class TagsAPITest(APITestCase):
             self.assertIn('text', suggestion)
             self.assertIn('score', suggestion)
             self.assertTrue(suggestion['text'].startswith('test'))
+    
+    def test_suggest_tags_api_with_unicode(self):
+        '''It should suggest tags'''
+        with self.autoindex():
+            for i in range(3):
+                tags = [faker.word(), faker.word(), 'testé',
+                        'testé-{0}'.format(i)]
+                ReuseFactory(tags=tags, visible=True)
+                DatasetFactory(tags=tags, visible=True)
+
+        response = self.get(url_for('api.suggest_tags'),
+                            qs={'q': 'testé', 'size': '5'})
+        self.assert200(response)
+
+        self.assertLessEqual(len(response.json), 5)
+        self.assertGreater(len(response.json), 1)
+        self.assertEqual(response.json[0]['text'], 'teste')
+
+        for suggestion in response.json:
+            self.assertIn('text', suggestion)
+            self.assertIn('score', suggestion)
+            self.assertTrue(suggestion['text'].startswith('teste'))
 
     def test_suggest_tags_api_no_match(self):
         '''It should not provide tag suggestion if no match'''